Incorporated review comments
